exp multiplier, get and change

+++ Hanekawa.Database/Tables/Config/Guild/LevelConfig.cs
@@ -3,7 +3,7 @@
     public class LevelConfig
     {
         public ulong GuildId { get; set; }
-        public int ExpMultiplier { get; set; } = 1;
+        public double ExpMultiplier { get; set; } = 1;
         public bool StackLvlRoles { get; set; } = true;
     }
 }
